==> R/00_Read_SDS_Attributes.R <==
farbe <- data.frame("farbe" = values(attr_hash[["farbe"]]), "Wert" = keys(attr_hash[["farbe"]]))
spalt_flaechen <- data.frame("spalt_flaechen" = values(attr_hash[["spalt_flaechen"]]), "Wert" = keys(attr_hash[["spalt_flaechen"]]))
strukturen_material <- data.frame("strukturen_material" = values(attr_hash[["strukturen_material"]]), "Wert" = keys(attr_hash[["strukturen_material"]]))
rinde <- data.frame("rinde" = values(attr_hash[["rinde"]]), "Wert" = keys(attr_hash[["rinde"]]))
sprung_flaechen <- data.frame("sprung_flaechen" = values(attr_hash[["sprung_flaechen"]]), "Wert" = keys(attr_hash[["sprung_flaechen"]]))
abrollung <- data.frame("abrollung" = values(attr_hash[["abrollung"]]), "Wert" = keys(attr_hash[["abrollung"]]))
therm_art <- data.frame("therm_art" = values(attr_hash[["therm_art"]]), "Wert" = keys(attr_hash[["therm_art"]]))
therm_zeit <- data.frame("therm_zeit" = values(attr_hash[["therm_zeit"]]), "Wert" = keys(attr_hash[["therm_zeit"]]))
erhaltung_gf <- data.frame("erhaltung_gf" = values(attr_hash[["erhaltung_gf"]]), "Wert" = keys(attr_hash[["erhaltung_gf"]]))
erhaltung_mod <- data.frame("erhaltung_mod" = values(attr_hash[["erhaltung_mod"]]), "Wert" = keys(attr_hash[["erhaltung_mod"]]))
gf_1 <- data.frame("gf_1" = values(attr_hash[["gf_1"]]), "Wert" = keys(attr_hash[["gf_1"]]))
erhaltung_gf_schlagrichtung <- data.frame("erhaltung_gf_schlagrichtung" = values(attr_hash[["erhaltung_gf_schlagrichtung"]]), "Wert" = keys(attr_hash[["erhaltung_gf_schlagrichtung"]]))
verlauf_schlagrichtung <- data.frame("verlauf_schlagrichtung" = values(attr_hash[["verlauf_schlagrichtung"]]), "Wert" = keys(attr_hash[["verlauf_schlagrichtung"]]))
schlagflaechenrest_erhaltung <- data.frame("schlagflaechenrest_erhaltung" = values(attr_hash[["schlagflaechenrest_erhaltung"]]), "Wert" = keys(attr_hash[["schlagflaechenrest_erhaltung"]]))
schlagflaechenrest_art <- data.frame("schlagflaechenrest_art" = values(attr_hash[["schlagflaechenrest_art"]]), "Wert" = keys(attr_hash[["schlagflaechenrest_art"]]))
schlagflaechenrest_form <- data.frame("schlagflaechenrest_form" = values(attr_hash[["schlagflaechenrest_form"]]), "Wert" = keys(attr_hash[["schlagflaechenrest_form"]]))
schlagmerkmal_schlagnarbe <- data.frame("schlagmerkmal_schlagnarbe" = values(attr_hash[["schlagmerkmal_schlagnarbe"]]), "Wert" = keys(attr_hash[["schlagmerkmal_schlagnarbe"]]))
schlagmerkmal_schlagaugen <- data.frame("schlagmerkmal_schlagaugen" = values(attr_hash[["schlagmerkmal_schlagaugen"]]), "Wert" = keys(attr_hash[["schlagmerkmal_schlagaugen"]]))
schlagmerkmal_kegel <- data.frame("schlagmerkmal_kegel" = values(attr_hash[["schlagmerkmal_kegel"]]), "Wert" = keys(attr_hash[["schlagmerkmal_kegel"]]))
menge_rinde <- data.frame("menge_rinde" = values(attr_hash[["menge_rinde"]]), "Wert" = keys(attr_hash[["menge_rinde"]]))
anzahl_negative_mit_bulbus <- data.frame("anzahl_negative_mit_bulbus" = values(attr_hash[["anzahl_negative_mit_bulbus"]]), "Wert" = keys(attr_hash[["anzahl_negative_mit_bulbus"]]))
anzahl_negative_ohne_bulbus <- data.frame("anzahl_negative_ohne_bulbus" = values(attr_hash[["anzahl_negative_ohne_bulbus"]]), "Wert" = keys(attr_hash[["anzahl_negative_ohne_bulbus"]]))
dorsal_praep <- data.frame("dorsal_praep" = values(attr_hash[["dorsal_praep"]]), "Wert" = keys(attr_hash[["dorsal_praep"]]))
praep_grat <- data.frame("praep_grat" = values(attr_hash[["praep_grat"]]), "Wert" = keys(attr_hash[["praep_grat"]]))
distalende_zustand <- data.frame("distalende_zustand" = values(attr_hash[["distalende_zustand"]]), "Wert" = keys(attr_hash[["distalende_zustand"]]))
abbaurichtung_dorsalflaechennegative <- data.frame("abbaurichtung_dorsalflaechennegative" = values(attr_hash[["abbaurichtung_dorsalflaechennegative"]]), "Wert" = keys(attr_hash[["abbaurichtung_dorsalflaechennegative"]]))
index_geraete_modifikation <- data.frame("index_geraete_modifikation" = values(attr_hash[["index_geraete_modifikation"]]), "Wert" = keys(attr_hash[["index_geraete_modifikation"]]))
rohmaterial <- data.frame("rohmaterial" = values(attr_hash[["rohmaterial"]]), "Wert" = keys(attr_hash[["rohmaterial"]]))
mod_art_dechsel_beil <- data.frame("mod_art_dechsel_beil" = values(attr_hash[["mod_art_dechsel_beil"]]), "Wert" = keys(attr_hash[["mod_art_dechsel_beil"]]))
mod_reihenfolge_dechsel_beil  <- data.frame("mod_reihenfolge_dechsel_beil" = values(attr_hash[["mod_reihenfolge_dechsel_beil"]]), "Wert" = keys(attr_hash[["mod_reihenfolge_dechsel_beil"]]))
